backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.4,<5.0
python-decouple>=3.8,<4.0
requests>=2.32,<3.0
